Made FilesystemWatcher private, renamed web_filesystem_watcher, cleaned up variables
